pyproject.toml
[project]
name = "mixedbread-parsing-haystack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "haystack-ai>=2.19.0",
    "mixedbread>=0.40.0",
    "tqdm>=4.67.1",
]
